+++ proto/cosmos/orm/v1alpha1/schema.proto
@@ -43,7 +43,7 @@ enum StorageType {
   // STORAGE_TYPE_DEFAULT_UNSPECIFIED indicates the persistent
   // KV-storage where primary key entries are stored in merkle-tree
   // backed commitment storage and indexes and seqs are stored in
-  // fast index storage. Note that the Cosmos SDK before store/v2
+  // fast index storage. Note that the Cosmos SDK before store/v2alpha1
   // does not support this.
   STORAGE_TYPE_DEFAULT_UNSPECIFIED = 0;
 
@@ -61,7 +61,7 @@ enum StorageType {
 
   // STORAGE_TYPE_INDEX indicates persistent storage which is not backed
   // by a merkle-tree and won't affect the app hash. Note that the Cosmos SDK
-  // before store/v2 does not support this.
+  // before store/v2alpha1 does not support this.
   STORAGE_TYPE_INDEX = 3;
 
   // STORAGE_TYPE_INDEX indicates persistent storage which is backed by
@@ -69,7 +69,7 @@ enum StorageType {
   // will affect the app hash and this is generally less efficient
   // than using STORAGE_TYPE_DEFAULT_UNSPECIFIED which separates index
   // keys into index storage. Note that modules built with the
-  // Cosmos SDK before store/v2 must specify STORAGE_TYPE_COMMITMENT
+  // Cosmos SDK before store/v2alpha1 must specify STORAGE_TYPE_COMMITMENT
   // instead of STORAGE_TYPE_DEFAULT_UNSPECIFIED or STORAGE_TYPE_INDEX
   // because this is the only type of persistent storage available.
   STORAGE_TYPE_COMMITMENT = 4;
