+++ store/v2alpha1/multi/store.go
@@ -17,10 +17,10 @@ import (
 	"github.com/cosmos/cosmos-sdk/store/listenkv"
 	"github.com/cosmos/cosmos-sdk/store/prefix"
 	"github.com/cosmos/cosmos-sdk/store/tracekv"
-	types "github.com/cosmos/cosmos-sdk/store/v2"
-	"github.com/cosmos/cosmos-sdk/store/v2/mem"
-	"github.com/cosmos/cosmos-sdk/store/v2/smt"
-	"github.com/cosmos/cosmos-sdk/store/v2/transient"
+	types "github.com/cosmos/cosmos-sdk/store/v2alpha1"
+	"github.com/cosmos/cosmos-sdk/store/v2alpha1/mem"
+	"github.com/cosmos/cosmos-sdk/store/v2alpha1/smt"
+	"github.com/cosmos/cosmos-sdk/store/v2alpha1/transient"
 	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
 	"github.com/cosmos/cosmos-sdk/types/kv"
 )
