+++ store/v2alpha1/smt/ics23_test.go
@@ -8,7 +8,7 @@ import (
 	"github.com/stretchr/testify/assert"
 
 	"github.com/cosmos/cosmos-sdk/db/memdb"
-	store "github.com/cosmos/cosmos-sdk/store/v2/smt"
+	store "github.com/cosmos/cosmos-sdk/store/v2alpha1/smt"
 )
 
 func TestProofICS23(t *testing.T) {
