+++ store/v2alpha1/smt/proof_test.go
@@ -8,7 +8,7 @@ import (
 	"github.com/stretchr/testify/require"
 
 	"github.com/cosmos/cosmos-sdk/db/memdb"
-	smtstore "github.com/cosmos/cosmos-sdk/store/v2/smt"
+	smtstore "github.com/cosmos/cosmos-sdk/store/v2alpha1/smt"
 	"github.com/lazyledger/smt"
 )
