+++ store/v2alpha1/smt/store_test.go
@@ -6,7 +6,7 @@ import (
 	"github.com/stretchr/testify/assert"
 
 	"github.com/cosmos/cosmos-sdk/db/memdb"
-	store "github.com/cosmos/cosmos-sdk/store/v2/smt"
+	store "github.com/cosmos/cosmos-sdk/store/v2alpha1/smt"
 )
 
 func TestGetSetHasDelete(t *testing.T) {
